asdf

+++ database/migrations/2024_09_25_195901_add_change_to_check_list_answers_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('check_list_answers', function (Blueprint $table) {
-            $table->tinyInteger('inspector_answered')->default(null)->change();
-            $table->tinyInteger('technic_answered')->default(null)->change();
-            $table->tinyInteger('author_answered')->default(null)->change();
+            $table->integer('inspector_answered')->default(null)->change();
+            $table->integer('technic_answered')->default(null)->change();
+            $table->integer('author_answered')->default(null)->change();
         });
     }
